requirements.txt
gradio
assemblyai
python-dotenv
pandas
openpyxl
openai
soundfile